ios 10

+++ Moobeez/Classes/Moobeez/View Controllers/MoviesViewController.m
@@ -334,7 +334,7 @@ typedef enum MoviesSection {
     // Leave the vertical pans for scrolling the content.
     if ([gestureRecognizer isKindOfClass:[UIPanGestureRecognizer class]]) {
         CGPoint translation = [gestureRecognizer translationInView:self.tableView];
-        BOOL isHorizontalPan = (fabsf(translation.x) > fabsf(translation.y));
+        BOOL isHorizontalPan = (fabs(translation.x) > fabs(translation.y));
         return  isHorizontalPan;
     }
